pyproject.toml
[project]
name = "categorizador-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Fabian",email = "heromfabian@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "categorizador_backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
